pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "todo_app_ssleimann"
version = "1.0"
authors = [
  { name="Sleiman Orocua", email="sleimanjose23@hotmail.com" },
]
description = "To-do API using Clean Architecture"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: FastAPI"
]
dependencies = [
  "fastapi[all]",
  "sqlalchemy",
  "sqlalchemy[asyncio]",
  "alembic",
  "aiosqlite",
  "pytest",
  "black",
  "python-jose[cryptography]",
  "passlib[bcrypt]",
  "cachetools"
]
keywords = ["clean architecture", "api", "to-do", "fastapi"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
